+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to Quantum will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.1.0]
+## [v1.3.0]
+
+### Fixed
+- Removed Unnecessary show() and raise() on url window open from web integration request.
+
+### Changed
+- Made quantum run in system tray instead of closing immediately.
+
+## [v1.2.0]
 
 ### Added
 - Added firefox integration initial files.